OHRM5X-2573: rename story
